Create BoardingGroup
Passengers are ordered by their surname's alphabetically, The people who are travelling together are placed to a group together and that boarding group is determined by the surname which is first alphabetically among the group. If a person's last name is in between a and f they belong to the boarding group one, if it is in between g and n they belong to boarding group two, if it is in o and t they belong to boarding group three, finally if it is between u and z they belong to boarding group four.